srcs/requirements/service_user_handler/django/conf/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.3.1
django-otp>=1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.6.0
gunicorn==23.0.0
h2==4.2.0
hpack==4.1.0
hvac==2.3.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
multidict==6.4.4
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
priority==1.3.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
qrcode>=7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.8.0
social-auth-app-django==5.4.2
social-auth-core==4.6.1
sqlparse==0.5.3
twilio>=9.3.0
Twisted==24.11.0
typing_extensions==4.13.0
urllib3==2.4.0
whitenoise==6.6.0
yarl==1.20.0
zope.interface==7.2
